Customize GTEST_PREFIX environment variable
By default, GTEST_PREFIX is set to /usr/src. However, sometimes it can
be installed in a different location. In this case, we can set the
GTEST_PREFIX environment variable to the correct location. If it is
unset, the default value will be used.

For example, in the lab machine of Coursrea class
https://www.coursera.org/specializations/gpu-programming, you can only
install gtest in /home/coder/project due to the lack of root permission.